Added ManageReputationContainer and updated SmiteDialog and AwardDialog
