+++ src/modules/dashboard/components/AwardAndSmiteDialogs/AwardAndSmiteDialogForm/index.ts
@@ -1,4 +1,3 @@
 export { default } from './AwardAndSmiteDialogForm';
 
 export * from './types';
-export { AwardAndSmiteFormValidationSchema } from './schemas';
