+++ src/modules/dashboard/components/AwardAndSmiteDialogs/AwardAndSmiteDialogForm/types.ts
@@ -1,3 +1,5 @@
+import { MessageDescriptor } from 'react-intl';
+
 import { WizardDialogType } from '~utils/hooks';
 import { DialogProps, ActionDialogProps } from '~core/Dialog';
 import { Address } from '~types/index';
@@ -14,5 +16,7 @@ export interface AwardAndSmiteDialogFormValues {
 export type AwardAndSmiteDialogProps = Required<DialogProps> &
   WizardDialogType<object> &
   ActionDialogProps & {
+    formMSG: Record<string, MessageDescriptor>;
+    isSmitingReputation: boolean;
     ethDomainId?: number;
   };
