+++ src/modules/dashboard/components/AwardAndSmiteDialogs/ManageReputationContainer/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ManageReputationContainer';
